HIVE-20531 : One of the task , either move or add partition can be avoided in repl load flow : test failure fix
